api/requirements.txt
fastapi
fastapi_cli
mypy
numpy
pandas
pandas-stubs
pydantic-yaml
ruff
sqlmodel
